Enhanced the packet parser by taking into account the pointer_field, based on Iuri's algorithm. We no longer have issues feeding PES packets, now.

+++ src/tables/nit.h
@@ -30,7 +30,7 @@ typedef struct nit_table {
 } __attribute__((__packed__)) nit_table;
 
 
-int nit_parse(const struct ts_header *header, const char *payload, uint8_t payload_len,
+int nit_parse(const struct ts_header *header, const char *payload, uint32_t payload_len,
 		struct demuxfs_data *priv);
 
 #endif /* __nit_h */
